Create Performance SASE.py
The code is a simple script that retrieves network performance data from a Cisco API endpoint and analyzes the data for potential issues. The script makes an HTTP GET request to the API endpoint using the requests library, passing along an api-key and access-token in the request headers. If the response status code is not 200 (i.e. success), the script raises an exception with an error message.

The response data, in JSON format, is loaded into a Python dictionary using the json.loads method. The script then checks the value of the network_latency key in the dictionary against a threshold value of 100 milliseconds. If the network_latency is greater than the threshold, the script outputs a message indicating a network latency issue and takes an action to resolve the issue (which is currently a placeholder in the form of a TODO comment).